+++ e3nn_jax/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.16.0"
+__version__ = "0.17.0"
 
 from e3nn_jax._src.config import config
 from e3nn_jax._src.rotation import (